Split up ApiPackageLookup to put the parts under test (Integration and unit) and to make it easier to do something else besides the globally latest version (e.g. only minor version update)
