+++ NuKeeper/NuGet/Api/IPackageVersionsLookup.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace NuKeeper.NuGet.Api
+{
+    public interface IPackageVersionsLookup
+    {
+        Task<IEnumerable<PackageSearchMedatadataWithSource>> Lookup(string packageName);
+    }
+}
